+++ commands.js
@@ -1,7 +1,8 @@
-const fs = require('fs');
+const {readdirSync} = require('fs');
+const {Collection} = require('discord.js');
 
-const commands = new Map();
-const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
+const commands = new Collection();
+const commandFiles = readdirSync('./commands').filter(file => file.endsWith('.js'));
 for(const file of commandFiles){
     const command = require(`./commands/${file}`);
     commands.set(command.name,command);