+++ commands/claim.js
@@ -1,10 +1,11 @@
-const Discord = require('discord.js');
+const {MessageEmbed} = require('discord.js');
 const Pet = require('../pets/pet.js');
 const locale = require('../locale/text.js');
 const petinfo = require('../pets/pets.json');
 
 module.exports = {
     name:'claim',
+    aliases:['c'],
     syntax:'/claim',
     admin:false,
     async execute(message,args,db){
@@ -20,15 +21,14 @@ module.exports = {
         userpets.stats.totalClaimed += 1;
         let newPet = new Pet({});
         let petstats = petinfo.pets[newPet.id];
+        let msg = new MessageEmbed().setTitle(`Pet claim: ${petstats.displayName}`);
         if(newPet.id in userpets.pets){
             const value = new Number(petstats.value) * userpets.tokens.multiplier;
             if(userpets.stats.petDuplicates == undefined) userpets.stats.petDuplicates = 0;
             userpets.stats.petDuplicates += 1;
             userpets.tokens.count += value;
             await db.petsdb.set(message.author.id,userpets);
-            let msg = new Discord.MessageEmbed()
-                .setTitle(`Pet claim: ${petstats.displayName}`)
-                .setColor('#FF0000')
+            msg.setColor('#FF0000')
                 .attachFiles([`./pets/sprites/${petstats.sprite}`])
                 .setImage(`attachment://${petstats.sprite}`)
                 .setDescription(`You have this pet in your inventory! As compensation, you received ${value} tokens.`)
@@ -39,9 +39,7 @@ module.exports = {
         if(userpets.stats.petsClaimed == undefined) userpets.stats.petsClaimed = 0;
         userpets.stats.petsClaimed += 1;
         await db.petsdb.set(message.author.id,userpets);
-        let msg = new Discord.MessageEmbed()
-            .setTitle(`Pet claim: ${petstats.displayName}`)
-            .setColor('#00FF00')
+        msg.setColor('#00FF00')
             .attachFiles([`./pets/sprites/${petstats.sprite}`])
             .setImage(`attachment://${petstats.sprite}`)
             .setDescription(`You unlocked ${petstats.displayName}! It was added to your collection.`)