+++ commands/source.js
@@ -1,12 +1,13 @@
-const Discord = require('discord.js');
+const {MessageEmbed} = require('discord.js');
 const locale = require('../locale/text.js');
 
 module.exports = {
     name:'source',
+    aliases:['code','github'],
     syntax:'/source',
     admin:false,
     execute(message,args,db){
-        let msg = new Discord.MessageEmbed()
+        let msg = new MessageEmbed()
             .setColor('#00FF00')
             .setTimestamp()
             .setTitle(locale.text({lang:db.lang,msg:"source_title"}))